Button: move component files into subfolder
