+++ src/lib/components/Button/index.ts
@@ -1,0 +1,3 @@
+export { default } from "./Button.svelte";
+
+export * from "./buttonOptions";
